requirements.txt
Pillow>=9.1.0
watchdog>=2.1.8
